+++ src/DataManager/ClientApp/minerva/ts/components/index.ts
@@ -1,1 +1,2 @@
 export * from './app-layout.component';
+export * from './require-auth.component';