+++ src/DataManager/ClientApp/minerva/ts/core/index.ts
@@ -1,3 +1,7 @@
+export * from './shared';
+export * from './http';
+export * from './user';
+
 export type NavItem = {
   navId: string,
   navName: string,