+++ src/DataManager/ClientApp/simulation/ts/core/index.ts
@@ -1,3 +1,7 @@
 import { NavItem } from 'minerva/core';
 
-export const navigator: NavItem[] = []
+export const navigator: NavItem[] = [{
+  navId: 'ads-manager',
+  navName: 'Ads Manager',
+  navPath: '/ads-manager'
+}]