+++ src/DataManager/ClientApp/simulation/ts/views/ads-manager.view.tsx
@@ -1,0 +1,9 @@
+export function AdsManagerView() {
+  return <>
+    <ol className="breadcrumb">
+      <li className="breadcrumb-item active">Ads Manager</li>
+    </ol>
+    <main className="fullscreen">
+    </main>
+  </>
+}