+++ src/DataManager/CHANGELOG.md
@@ -5,6 +5,9 @@
 **Features**
 
 - Build client app with bootstrap, react, rosie-ui
+- Add OAuth2
+- Add `Simulation` tab
+  - Add `Ads Manager` screen
 
 ## v0.0.1
 > 2025-12-07